requirements.txt
hvac
azure-storage-blob
PyYAML